Revert "Core splitting (#181)"
This reverts commit bac9e942a578f30f5d21d0c70863a8a353fdb596.
