+++ wfsim/__init__.py
@@ -1,15 +1,5 @@
 __version__ = "0.5.4"
 
-from . import core
 from .core import *
-from .core.rawdata import *
-from .core.pulse import *
-from .core.s1 import *
-from .core.s2 import *
-from .core.afterpulse import *
-
 from .strax_interface import *
 from .pax_interface import *
-
-from .load_resource import *
-from .utils import *
